Reorganize tests to make testing easier
