+++ packages/liveblocks-react/jest.config.js
@@ -1,5 +1,6 @@
 // jest.config.js
 module.exports = {
   modulePathIgnorePatterns: ["<rootDir>/lib/"],
+  testPathIgnorePatterns: ["__tests__/_.*"],
   setupFiles: ["./jest.setup.js"],
 };
