+++ packages/liveblocks-react/src/__tests__/_liveblocks.config.ts
@@ -1,0 +1,19 @@
+import type { LiveObject } from "@liveblocks/client";
+import { createClient } from "@liveblocks/client";
+
+import { createRoomContext } from "../factory";
+
+type Presence = {
+  x: number;
+};
+
+type Storage = {
+  obj: LiveObject<{
+    a: number;
+  }>;
+};
+
+const client = createClient({ authEndpoint: "/api/auth" });
+
+export const { RoomProvider, useMyPresence, useObject, useOthers } =
+  createRoomContext<Presence, Storage>(client);
